pyproject.toml
[tool.poetry]
name = "tf-playwright-stealth"
packages = [{ include = "playwright_stealth" }]
version = "1.2.0"
description = "Makes playwright stealthy like a ninja!"
authors = []
homepage = "https://www.agentql.com/"
repository = "https://github.com/tinyfish-io/tf-playwright-stealth"
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
playwright = "^1"
fake-http-header = "^0.3.5"

[tool.poetry.group.dev.dependencies]
agentql = "^1.3.0"
pytest = "^8.3.3"
pytest-mockito = "^0.0.4"
dill = "*"
ruff = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true
docstring-code-line-length = 20

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
exclude = ["script/**"]
